backend/requirements.txt
# ── Core framework ──────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# ── LangGraph / LangChain ───────────────────────────────────────────────────
langgraph==0.2.62
langchain-core==0.3.29
langchain-groq==0.2.3

# ── LLM / structured output ─────────────────────────────────────────────────
groq==0.13.1
instructor==1.7.2

# ── Database ─────────────────────────────────────────────────────────────────
asyncpg==0.30.0
sqlalchemy==2.0.36          # for type helpers; not ORM

# ── Supabase client ──────────────────────────────────────────────────────────
supabase==2.28.3

# ── Auth / JWT ───────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
PyJWT==2.10.1

# ── HTTP ─────────────────────────────────────────────────────────────────────
httpx==0.28.1
python-multipart==0.0.20    # for file upload endpoints

# ── Observability ────────────────────────────────────────────────────────────
structlog==25.1.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0

# ── LangSmith (optional tracing) ─────────────────────────────────────────────
langsmith==0.2.10

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==9.0.0
orjson==3.10.12
anyio==4.8.0

# ── Dev / test (install separately in CI with pip install -r requirements-dev.txt) ──
# see requirements-dev.txt
